\name{sample_data}
\alias{sample_data}
\docType{data}
\title{
'sample data'
}
\description{
An example for the data input.
}
\usage{data("sample_data")}
\format{
  A data frame with 22000 observations on the following 4 variables.
  \describe{
    \item{\code{V1}}{a factor with levels}
    \item{\code{V2}}{a numeric vector}
    \item{\code{V3}}{a numeric vector}
    \item{\code{V4}}{a numeric vector}
  }
}
\details{
The input data must be same as the sample data.
}
\source{
nothing
}
\references{
2016 tutorial
}
\examples{
data(sample_data)
## maybe str(sample_data) ; plot(sample_data) ...
}
\keyword{datasets}
